% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redshift.R
\name{generate_sql_load}
\alias{generate_sql_load}
\title{Generate SQL loading statements from a manifest file. This assumes that the
tables need to be created and then the data copied in.}
\usage{
generate_sql_load(manifest_file, s3_url)
}
\arguments{
\item{manifest_file}{local path to the manifest file}

\item{s3_url}{the S3 bucket where the manifest will be stored for loading}
}
\value{
nothing; SQL statements written to file
}
\description{
Generate SQL loading statements from a manifest file. This assumes that the
tables need to be created and then the data copied in.
}
