% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redshift.R
\name{genereate_stata_infile}
\alias{genereate_stata_infile}
\title{Create a Stata infile script}
\usage{
genereate_stata_infile(manifest_file)
}
\arguments{
\item{manifest}{the manifest to be processed}
}
\value{
nothing; file is written directly
}
\description{
Create a Stata infile script
}
