% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redshift.R
\name{insert_header}
\alias{insert_header}
\title{Create scripts to add header lines to all files in a manifest}
\usage{
insert_header(manifest_file)
}
\arguments{
\item{manifest_file}{the manifest to be processed}
}
\value{
silent; the script is written directly
}
\description{
Create scripts to add header lines to all files in a manifest
}
