% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_manifest.R
\name{manifest_filelist}
\alias{manifest_filelist}
\title{Return a list of the data files referenced by this manifest}
\usage{
manifest_filelist(manifest)
}
\arguments{
\item{manifest}{path to a manifest file collocated with data}
}
\value{
vector of file names
}
\description{
Return a list of the data files referenced by this manifest
}
