% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_manifest.R
\name{manifest_headers}
\alias{manifest_headers}
\title{Extract the header names for the data}
\usage{
manifest_headers(manifest)
}
\arguments{
\item{manifest}{local path to the manifest file}
}
\value{
vector of header names, in the order in which they appear; vector of
column types, in the same order
}
\description{
Extract the header names for the data
}
