% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_manifest.R
\name{manifest_records}
\alias{manifest_records}
\title{Extract the total number of records according to the manifest}
\usage{
manifest_records(manifest)
}
\arguments{
\item{manifest}{local path to the manifest file}
}
\value{
integer of the total number of records in the manifest
}
\description{
Extract the total number of records according to the manifest
}
