% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRedshift.R
\name{read_redshift}
\alias{read_redshift}
\title{read_redshift will read in the contents of a manifest and return a data.frame}
\usage{
read_redshift(manifest_file)
}
\arguments{
\item{manifest_file}{path to a Redshift manifest file}
}
\value{
the data.frame containing the validated data specified
by the manifest file. Delimiters will be detected automatically unless one
is specified.
}
\description{
read_redshift will read in the contents of a manifest and return a data.frame
}
\examples{
venue_data <- read_redshift(system.file("extdata","venuemanifest",
 package="ReadRedshift"))
}
