% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRedshift.R
\name{readRedshift}
\alias{readRedshift}
\title{readRedshift will read in the contents of a manifest and return a data.frame}
\usage{
readRedshift(manifest_file)
}
\arguments{
\item{manifest_file}{path to a RedShift manifest file}
}
\value{
the data.frame containing the validated data specified
by the manifest file. Delimiters will be detected automatically unless one
is specified.
}
\description{
readRedshift will read in the contents of a manifest and return a data.frame
}
