% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{redshift_to_r_data}
\alias{redshift_to_r_data}
\title{A list to convert from RedShift data types to readr column specs}
\format{
An object of class \code{list} of length 4.
}
\usage{
redshift_to_r_data
}
\description{
A list to convert from RedShift data types to readr column specs
}
\keyword{datasets}
