% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redshift.R
\name{redshift_to_sql_data}
\alias{redshift_to_sql_data}
\title{Internal function to convert data, since SQL is harder to handle}
\usage{
redshift_to_sql_data(type_list)
}
\arguments{
\item{type_list}{type list with parameters for single data element}
}
\value{
a string representing this type in SQL format
}
\description{
Internal function to convert data, since SQL is harder to handle
}
\examples{
redshift_to_sql_data(
list("base"= "character varying", "max_length"= 25))
}
