% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{venue}
\alias{venue}
\title{Simple dataset used for testing read_redshift functions}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 6 rows and 5 columns.
}
\source{
\url{https://docs.aws.amazon.com/redshift/latest/dg/}

Last accessed on Jan 25, 2022.
}
\usage{
venue
}
\description{
This dataset, and the manifests used to read it, were copied from the
UNLOAD examples in the RedShift documentation.
}
\keyword{datasets}
